refactor project structure
in preparation for the big restructure using data classes:

* rename some files and folders
* update Readmes / restructure Readmes
* move argument parser into automatic_walk_time_tables modules
