+++ .gitignore
@@ -1,4 +1,4 @@
 python_program/.idea
 .idea
-python_program/res/swisstlm3d_2021-04_2056_5728.gdb
-python_program/res/swissnames3d_2021_2056.shp
+python_program/automatic_walk_time_tables/res/swisstlm3d_2021-04_2056_5728.gdb
+python_program/automatic_walk_time_tables/res/swissnames3d_2021_2056.shp
