+++ python_program/.dockerignore
@@ -1,1 +1,1 @@
-**/res/original_data
+**/automatic_walk_time_tables/res/original_data
