+++ python_program/automatic_walk_time_tables/geo_processing/find_swisstopo_name.py
@@ -1,5 +1,6 @@
 import csv
 import logging
+import os
 from math import sqrt
 
 logger = logging.getLogger(__name__)
@@ -78,11 +79,13 @@ def find_name(coord, dist):
 
 database = []
 
+dirname = os.path.dirname(__file__)
+
 # Linien (Verkehrsbauten, Sportanlagen, Fliessgewässern ...)
-add_to_database('./res/swissNAMES3D_LIN.csv', database, 1, 5, 10, 11)
+add_to_database(os.path.join(dirname, '../res/swissNAMES3D_LIN.csv'), database, 1, 5, 10, 11)
 
 # Punkte (Topografische Namen, Flur- und Lokalnamen, Gebäudenamen ...)
-add_to_database('./res/swissNAMES3D_PKT.csv', database, 1, 6, 11, 12)
+add_to_database(os.path.join(dirname, '../res/swissNAMES3D_PKT.csv'), database, 1, 6, 11, 12)
 
 # Polygone (Siedlungsnamen, Seenamen, Geländenamen ..)
-add_to_database('./res/swissNAMES3D_PLY.csv', database, 1, 5, 10, 11)
+add_to_database(os.path.join(dirname, '../res/swissNAMES3D_PLY.csv'), database, 1, 5, 10, 11)
