+++ python_program/automatic_walk_time_tables/res/swissNAMES3D_LIN.csv
